Add files via upload
Analysis of 10,051 Walmart branch transactions across 100 locations using Python and SQL uncovered several key business Insights.
Most notably, it found that digital payments dominated with 81.3% of 
transactions and that food and beverages and health and beauty were the top
performing categories with ratings of 7.11 and 7.00 out of 10, respectively. 
There is also potential for growth in the consumer electronics and home 
lifestyle sectors.